Pre locate the git executable in vcs git blocking
I have seen some indication in a bench that this may help loops that
run [vcs] process over and over. That bench wasn't precisely
reproducible, but I'd like to get this part of the execution out of
the way to focus on other parts.

I didn't do the same change in the eio/ part of the library because I
don't need it at the moment.