Remove the CLI help menu from the cram tests
The exact formatting of this output may change with dependencies upgrades and
it's best to rely on other tests to exercise the code.

Besides, we already monitor this menu in the documentation, where we can be a
bit more permissive as we'd expect to compile the doc with exactly one specific
combination of dependencies.